Add files via upload
This file is the cumulative analysis of our forecast for training hours for Mach 1 Flight Training center.  In this analysis, we attempt to forecast using time series forecast methods such as MA, Naive, and others.  We also create a multiple regression model using economic indicators.  The analysis yields promising results, though we simplify the model using a stepwise analysis.  Finally, we look to find the strongest model using the ARIMA analysis.